first draft - add dotnet debugging

+++ src/debug/debugSession.ts
@@ -5,7 +5,7 @@ import * as path from "path";
 import * as portfinder from "portfinder";
 import { ChildProcess } from "child_process";
 
-import { IDebugProvider } from "./debugProvider";
+import { IDebugProvider, PortInfo } from "./debugProvider";
 import * as providerRegistry from "./providerRegistry";
 
 import { kubeChannel } from "../kubeChannel";
@@ -115,7 +115,7 @@ export class DebugSession implements IDebugSession {
 
                 // Start debug session.
                 p.report({ message: `Starting ${this.debugProvider!.getDebuggerType()} debug session...`});  // safe because checked outside the lambda
-                await this.startDebugSession(appName, cwd, proxyResult);
+                await this.startDebugSession(appName, cwd, proxyResult, podName);
             } catch (error) {
                 vscode.window.showErrorMessage(error);
                 kubeChannel.showOutput(`Debug on Kubernetes failed. The errors were: ${error}.`);
@@ -159,25 +159,34 @@ export class DebugSession implements IDebugSession {
         }
 
         // Find the debug port to attach.
-        const portInfo = await this.debugProvider.resolvePortsFromContainer(this.kubectl, targetPod, targetPodNS, targetContainer);
-        if (!portInfo || !portInfo.debugPort) {
-            await this.openInBrowser("Cannot resolve the debug port to attach. See the documentation for how to use this command.", debugCommandDocumentationUrl);
-            return;
+        const isPortRequired = this.debugProvider.isPortRequired();
+        var portInfo: PortInfo | undefined = undefined;
+        if (isPortRequired) {
+            portInfo = await this.debugProvider.resolvePortsFromContainer(this.kubectl, targetPod, targetPodNS, targetContainer);
+            if (!portInfo || !portInfo.debugPort) {
+                await this.openInBrowser("Cannot resolve the debug port to attach. See the documentation for how to use this command.", debugCommandDocumentationUrl);
+                return;
+            }
         }
 
         vscode.window.withProgress({ location: vscode.ProgressLocation.Window }, async (p) => {
             try {
-                // Setup port-forward.
-                p.report({ message: "Setting up port forwarding..."});
-                const proxyResult = await this.setupPortForward(targetPod, targetPodNS, portInfo.debugPort);
+                var proxyResult: ProxyResult | undefined;
+                if (isPortRequired) {
+                    p.report({ message: "Setting up port forwarding..."});
+                    if (portInfo) {
+                        proxyResult = await this.setupPortForward(targetPod, targetPodNS, portInfo.debugPort);
+                    }
 
-                if (!proxyResult.proxyProcess) {
-                    return;  // No port forwarding, so can't debug
+                    if (!proxyResult || !proxyResult.proxyProcess) {
+                        return;  // No port forwarding, so can't debug
+                    }
                 }
 
                 // Start debug session.
                 p.report({ message: `Starting ${this.debugProvider!.getDebuggerType()} debug session...`});  // safe because checked outside lambda
-                await this.startDebugSession(undefined, workspaceFolder.uri.fsPath, proxyResult);
+
+                await this.startDebugSession(undefined, workspaceFolder.uri.fsPath, proxyResult, targetPod);
             } catch (error) {
                 vscode.window.showErrorMessage(error);
                 kubeChannel.showOutput(`Debug on Kubernetes failed. The errors were: ${error}.`);
@@ -326,11 +335,15 @@ export class DebugSession implements IDebugSession {
         return proxyResult;
     }
 
-    private async startDebugSession(appName: string | undefined, cwd: string, proxyResult: ProxyResult): Promise<void> {
+    private async startDebugSession(appName: string | undefined, cwd: string, proxyResult: ProxyResult | undefined, pod: string): Promise<void> {
         kubeChannel.showOutput("Starting debug session...", "Start debug session");
         const sessionName = appName || `${Date.now()}`;
-        await this.startDebugging(cwd, sessionName, proxyResult.proxyDebugPort, proxyResult.proxyAppPort, async () => {
-            if (proxyResult.proxyProcess) {
+
+        const proxyDebugPort = proxyResult ? proxyResult.proxyDebugPort : undefined;
+        const proxyAppPort = proxyResult ? proxyResult.proxyAppPort : undefined;
+
+        await this.startDebugging(cwd, sessionName, proxyDebugPort, proxyAppPort, pod, async () => {
+            if (proxyResult && proxyResult.proxyProcess) {
                 proxyResult.proxyProcess.kill();
             }
             if (appName) {
@@ -421,7 +434,7 @@ export class DebugSession implements IDebugSession {
         return forwardingRegExp.test(message);
     }
 
-    private async startDebugging(workspaceFolder: string, sessionName: string, proxyDebugPort: number, proxyAppPort: number, onTerminateCallback: () => Promise<any>): Promise<boolean> {
+    private async startDebugging(workspaceFolder: string, sessionName: string, proxyDebugPort: number | undefined, proxyAppPort: number | undefined, pod: string, onTerminateCallback: () => Promise<any>): Promise<boolean> {
         const disposables: vscode.Disposable[] = [];
         disposables.push(vscode.debug.onDidStartDebugSession((debugSession) => {
             if (debugSession.name === sessionName) {
@@ -444,7 +457,7 @@ export class DebugSession implements IDebugSession {
             return false;
         }
 
-        const success = await this.debugProvider.startDebugging(workspaceFolder, sessionName, proxyDebugPort);
+        const success = await this.debugProvider.startDebugging(workspaceFolder, sessionName, proxyDebugPort, pod);
         if (!success) {
             disposables.forEach((d) => d.dispose());
             await onTerminateCallback();
